Move extensions into newly-created companion objects
Where the receiver type lacked a companion object, create one and move
the extension there.

- `contingency.Mitigation.within`, `Recovery.within`, `Accrue.within`,
  `Tracking.within`, `Validate.within` → newly-created companions
- `probably.Test.{aspire, assert, check, matches}` → new `object Test`
- `vicarious.Catalog.brush` → new `object Catalog`
