+++ lib/contingency/src/core/contingency.Tracking.scala
@@ -32,8 +32,21 @@
                                                                                                   */
 package contingency
 
+import fulminate.*
 import proscenium.*
 import vacuous.*
+
+object Tracking:
+  extension [accrual <: Exception, lambda[_], focus]
+            (inline track: Tracking[accrual, lambda, focus])
+    inline def within[result](inline lambda: Foci[focus] ?=> lambda[result])
+      ( using tactic: Tactic[accrual], diagnostics: Diagnostics )
+    :   result =
+
+      $ {
+          contingency.internal.trackWithin[accrual, lambda, result, focus]
+            ( 'track, 'lambda, 'tactic, 'diagnostics )
+        }
 
 class Tracking[accrual <: Exception, lambda[_], focus]
   ( val initial: accrual,
