+++ lib/probably/src/core/probably_core.scala
@@ -111,22 +111,6 @@ def suite[report](name: Message)(using suite: Testable, runner: Runner[report])
   runner.suite(Testable(name, suite), block)
 
 
-extension [test](test: Test[test])
-  inline def aspire(inline predicate: test => Boolean): Unit =
-    ${probably.internal.aspire[test]('test)}
-
-
-  inline def assert(inline predicate: test => Boolean): Unit =
-    ${probably.internal.assert[test]('test, 'predicate)}
-
-
-  inline def check(inline predicate: test => Boolean): test =
-    ${probably.internal.check[test]('test, 'predicate)}
-
-  inline def assert(): Unit = ${probably.internal.assert[test]('test, '{probably.internal.succeed})}
-  inline def check(): test = ${probably.internal.check[test]('test, '{probably.internal.succeed})}
-  inline def matches(inline pf: test ~> Any): Unit = assert(pf.isDefinedAt(_))
-
 extension [value](inline value: value)(using inline test: Harness)
   inline def debug: value = ${probably.internal.debug('value, 'test)}
