+++ lib/probably/src/core/soundness_probably_core.scala
@@ -36,7 +36,7 @@ export
   probably
   . { Baseline, Benchmark, Inclusion, Verdict, Runner, Test, Harness, TestId, Report, Reporter,
       Trial, Testable, Tolerance, Min, Mean, Max, Cadential, ===, !==, Temporal, Geometric, Arithmetic,
-      +/-, ±, test, suite, aspire, assert, check, matches, debug, Checkable, Ci, GithubActions }
+      +/-, ±, test, suite, debug, Checkable, Ci, GithubActions }
 
 package harnesses:
   export probably.harnesses.threadLocal
