+++ lib/vicarious/src/core/vicarious_core.scala
@@ -45,13 +45,3 @@ inline def catalog[key](key: key)[value](inline lambda: [field] => (field: field
   ${vicarious.internal.catalog[key, value]('lambda, 'key, 'classTag)}
 
 
-extension [key, value: ClassTag](catalog: Catalog[key, value])
-  def brush(using proxy: Proxy[key, value, Nat])
-    ( lambda: (`*`: proxy.type) ?=> Proxy[key, value, Nat] ~> value )
-  :   Catalog[key, value] =
-
-    val partialFunction = lambda(using proxy)
-
-    Catalog(IArray.tabulate(catalog.size): index =>
-      partialFunction.applyOrElse
-        ( Proxy[key, value, index.type](), _ => catalog.values(index) ))
